benv: change default map from dvorak => qwerty
This alleviates confusion when the Windows/X keyboard map is set to dvorak, as
is the usual case on my machines. When a dvorak mapping is active at the OS level,
key presses on a qwerty keyboard are translated to the corresponding dvorak
characters. However, when my ErgoDox keyboard does the translation on its
own (i.e. generates Dvorak characters directly), the second OS-level translation
still takes place and produces seemingly random characters.