benv: map Tab key -> ' and Enter key -> L
The top row pinky keys ("'" and "L" in Dvorak) are too hard to reach
on the Ergodox. Map those characters to the outside keys of the home
row instead (Tab key on the left and Enter key on the right).

This makes typing on the Ergodox *much* more comfortable for me.