_STENO: add dual-function symbol/Escape key on CapsLock
Implement a dual-function key that acts Escape when pressed alone, or
as the chord prefix for Emily's Symbols (`KWH` in my case) when
pressed in combination with other keys.

This was tricky to implement, but it allows me to type
symbols (e.g. `!`, `@`) and to press Escape _much_ more easily, both
of which I do a lot!