_STENO: remap right `*` key to `TSDZ` (for `ben-fingerspelling.py`)
Implement custom behavior on the lower right `*` key (i.e. `STN_ST4`),
for use with `ben-fingerspelling.py` (my custom fingerspelling
dictionary).

When pressed alone, the `STN_ST4` key behaves like a normal `*`
key (i.e. it undoes the previous steno stroke), but when pressed in
combination with other keys, `STN_ST4` remaps to `TSDZ`, which is my
"unique ender" chord for `ben-fingerspelling.py`.

I wrote `ben-fingerspelling.py` to augment the standard fingerspelling
chords with spacing control, so that a space can be added before
and/or after the typed letter with the E/U keys. It is very similar to
the way that spacing/attachment is controlled with the Emily's Symbols
dictionary. `ben-fingerspelling.py` also adds new left-hand chords for
the letters e/u/i, since E/U keys are now used for spacing instead. As
a result, *all* letters can be fingerspelled with left-hand chords,
which I expect will be faster and more comfortable than the standard
fingerspelling chords.